Function bevy_egui_kbgp::kbgp_prepare

pub fn kbgp_prepare(
    egui_ctx: &Context, 
    prepare_dlg: impl FnOnce(KbgpPrepare<'_>)
)

Must be called every frame, either manually or by using KbgpPlugin.

Should be called between bevy_egui’s input handling system and the system that generates the UI - so in the CoreStage::PreUpdate stage after the EguiSystem::ProcessInput label.

The prepare_dlg argument is a closure that accepts a KbgpPrepare, and used to:

Register the input from the keyboard and the gamepads.

Set preferences.

Typical usage:

fn custom_kbgp_system(
    mut egui_context: ResMut<EguiContext>,
    keys: Res<Input<KeyCode>>,
    gamepads: Res<Gamepads>,
    gamepad_axes: Res<Axis<GamepadAxis>>,
    gamepad_buttons: Res<Input<GamepadButton>>,
    mouse_buttons: Res<Input<MouseButton>>,
    settings: Res<KbgpSettings>,
) {
    kbgp_prepare(egui_context.ctx_mut(), |prp| {
        match prp {
            KbgpPrepare::Navigation(prp) => {
                prp.navigate_keyboard_by_binding(&keys, &settings.bindings.keyboard);
                prp.navigate_gamepad_by_binding(&gamepads, &gamepad_axes, &gamepad_buttons, &settings.bindings.gamepad_buttons);
            }
            KbgpPrepare::PendingInput(prp) => {
                prp.accept_keyboard_input(&keys);
                prp.accept_mouse_buttons_input(&mouse_buttons);
                prp.accept_gamepad_input(&gamepads, &gamepad_axes, &gamepad_buttons);
            }
        }
    });
}