Struct bevy_egui_kbgp::KbgpPreparePendingInput

pub struct KbgpPreparePendingInput { /* private fields */ }

An option of KbgpPrepare.

Implementations

impl KbgpPreparePendingInput

pub fn accept_input(&mut self, input: KbgpInput)

Notify KBGP about a single input was accepted from the player.

pub fn accept_input_unique(&mut self, input: KbgpInput)

Notify KBGP about a single input was accepted from the player, only if the same input was not already received this frame.

pub fn accept_inputs(&mut self, inputs: impl Iterator<Item = KbgpInput>)

Notify KBGP about multiple inputs accepted from the player.

pub fn accept_keyboard_input(&mut self, keys: &Input<KeyCode>)

Notify KBGP about all the input from the keyboard.

pub fn accept_mouse_buttons_input(&mut self, buttons: &Input<MouseButton>)

Notify KBGP about all the input from mouse buttons.

pub fn accept_mouse_wheel_event(
&mut self,
event: &MouseWheel,
vertical: bool,
horizontal: bool
)

Notify KBGP about all the input from the mouse wheel.

pub fn accept_gamepad_input(
&mut self,
gamepads: &Gamepads,
axes: &Axis<GamepadAxis>,
buttons: &Input<GamepadButton>
)

Notify KBGP about all the input from the gamepad.